appletb: Add support for dimming and waking via touch.
Switch to using interface 3 to dim and turn off the touchbar display.
Because this only turns off the display, not the complete touchbar, we
still get touch events from it, and hence touches on the touchbar will
now wake the display just like keyboard or touchpad events.

Dimming is now performed after 2/3 of the idle-timeout or 30 seconds,
whichever is smaller.

Power measurements performed by looking at the battery drain current (as
reported by the battery) indicated that turning off just the display
saved about 25mW, with the dimmed state somewhere around half; enabling
USB suspend on the touchbar device, however, saves about 200mW - 250mW.
For this reason we enable USB suspend whenever turning off the display.
But there is a bit of a USB control-message storm whenever the touchbar
is touched while the device is suspened (~15K control message requests
in 50ms).

One outstanding issue with this change is that we don't properly create
a hid-device for the interface 3, but instead just blindly access it.